Small inconsistency: Removed generic Intent i and moved dependencies to using explicit intents
As commented by : @nelenkov
"Thanks for the patch. However, I don't think this makes it more readable. Getting rid of the Intent i would be better IMHO. It's not like the switch just sets up the intent and it is started at the end of the method -- some cases use startActivity() and others startActivityForResult() so it is better each case to use an explicit, named intent."